Create grants-ingest library
